requirements.txt
absl-py==2.3.0
accelerate==1.8.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==5.0.1
attrs==25.3.0
av==14.4.0
bitsandbytes==0.41.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
datasets==4.0.0
decorator==5.2.1
decord==0.6.0
deepspeed==0.14.4
dill==0.3.7
docstring_parser==0.16
einops @ file:///home/conda/feedstock_root/build_artifacts/einops_1739083394647/work
einops-exts==0.0.4
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.13
filelock==3.18.0
flash-attention==1.0.0
flash_attn==2.7.4.post1
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2023.10.0
ftfy==6.3.1
git-filter-repo==2.47.0
gitdb==4.0.12
GitPython==3.1.44
gmpy2 @ file:///home/conda/feedstock_root/build_artifacts/gmpy2_1745509363766/work
gradio_client==0.2.9
grpcio==1.73.0
h11==0.14.0
hf-xet==1.1.4
hf_transfer==0.1.9
hjson==3.1.0
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.33.0
idna==3.10
ipdb==0.13.13
ipython==8.37.0
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
kornia==0.8.1
kornia_rs==0.1.9
latex2mathml==3.78.0
-e git+https://github.com/LLaVA-VL/LLaVA-NeXT.git@b42941ceba259d5df18f8df8193a3897296a0449#egg=llava
Markdown==3.8.2
markdown-it-py==3.0.0
markdown2==2.5.3
MarkupSafe==3.0.2
matplotlib==3.10.3
matplotlib-inline==0.1.7
mdurl==0.1.2
modelscope==1.27.1
mpmath==1.3.0
multidict==6.5.0
multiprocess==0.70.15
networkx==3.4.2
ninja==1.11.1.4
numpy @ file:///home/conda/feedstock_root/build_artifacts/numpy_1747544640217/work/dist/numpy-2.2.6-cp310-cp310-linux_x86_64.whl#sha256=d6d964caeef85d00073d27cd62b46883d275b3d8162f723f0fcabbd0b3cc3f9d
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-ml-py==12.575.51
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
open_clip_torch==2.32.0
opencv-python==4.11.0.86
optree @ file:///home/conda/feedstock_root/build_artifacts/optree_1748442447565/work
packaging==25.0
pandas==2.3.0
parso==0.8.4
peft==0.4.0
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
prompt_toolkit==3.0.51
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==20.0.0
pyarrow-hotfix==0.7
pybind11 @ file:///D:/bld/pybind11-split_1747934313674/work
pybind11_global @ file:///home/conda/feedstock_root/build_artifacts/pybind11-split_1747934270866/work
pydantic==1.10.8
Pygments==2.19.1
pynvml==12.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
rich==14.0.0
rouge==1.0.1
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
sentencepiece==0.1.99
sentry-sdk==2.30.0
setproctitle==1.3.6
shortuuid==1.0.13
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
stack-data==0.6.3
starlette==0.46.2
svgwrite==1.4.3
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
threadpoolctl==3.6.0
timm==1.0.15
tokenizers==0.19.1
tomli==2.2.1
torch==2.6.0+cu126
torch-tb-profiler==0.4.3
torchaudio==2.6.0+cu126
torchvision==0.21.0+cu126
tqdm==4.67.1
traitlets==5.14.3
transformers==4.44.0
triton==3.2.0
typeguard==4.4.4
typing_extensions==4.14.1
tyro==0.9.24
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.34.3
wandb==0.20.1
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==15.0.1
Werkzeug==3.1.3
xxhash==3.5.0
yarl==1.20.1
